Split numbers cldr_serde
